FIX. fixed category vs. dynamic data
options.categories contains a fixed list of categories for the resulting items loaded vía url (dynamic).
however, when resulting data does not contain items for one of the categories an error is thrown. 
with the following check, no listBuilder is generated for those categories without resulting items for the autocomplete.